backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
python-multipart==0.0.18
yfinance==0.2.51
pytz==2024.2
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
ruff==0.8.6
